Move similarity config onto model
The previous API required callers to specify which
association to use for computing similarity on a given
record. This change moves the responsibility for specifying
an association into the model itself.

This change has a couple benefits:

* centralizes configuration of similarity rules onto the model.
  This will be beneficial in allowing for more complex config
  (e.g. weighted combinations of different associations) w/o
  requiring each caller to re-specify the similarity rules
* Even for simpler configurations, results in a cleaner API

One drawback is that we no longer support multiple types of
similarity.
In many cases, support for a single type of similarity is
probably adequate. But we can move back to supporting multiple
similarity types in the future by just allowing multiple
calls to similar_by which assign names to the different
configurations.